pyproject.toml
[project]
name = "healthcare-digital-fte"
version = "0.1.0"
description = "AI system automating medical coding, CDI, and revenue cycle workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3.0",
    "structlog>=24.0",
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pyjwt>=2.12.1",
    "cryptography>=46.0.6",
    "google-genai>=1.73.0",
    "groq>=1.1.2",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "compliance: marks compliance guardrail tests (run with -m compliance)",
    "hard_guardrail: marks hard guardrail tests (P0 priority)",
    "soft_guardrail: marks soft guardrail tests",
    "mimic: marks MIMIC-IV accuracy tests (requires downloaded data; skip with -m 'not mimic')",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
